Split util package into parser and storage packages
