+++ src/main/java/duke/command/ListCommand.java
@@ -2,8 +2,8 @@ package duke.command;
 
 import java.util.function.Consumer;
 
+import duke.storage.Storage;
 import duke.task.TaskList;
-import duke.util.Storage;
 
 /**
  * Encapsulates a command to list {@code Task}s in a {@code TaskList}.
