+++ src/main/java/duke/parser/Parser.java
@@ -1,4 +1,4 @@
-package duke.util;
+package duke.parser;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
